moved jobs functions to their own file
