+++ changelog/0.4.7.md
@@ -4,6 +4,7 @@
 
 - The `upload_notebook_file` function now includes a step to transform the notebook content to a valid format before validation and upload.
 - `get_user_id` tool is now called `get_user_info` and retrives all the information about the user.
+- New `create_schedule_jobs_from_notebook` tool to create schedule jobs from a notebook file.
 
 ## Fixed
