+++ src/api/tools/jobs/__init__.py
@@ -1,0 +1,11 @@
+"""Notebook tools for SingleStore MCP server."""
+
+from .jobs import (
+    create_job_from_notebook,
+    delete_job,
+)
+
+__all__ = [
+    "create_job_from_notebook",
+    "delete_job",
+]
