+++ src/api/tools/notebooks/__init__.py
@@ -3,11 +3,6 @@
 from .notebooks import (
     create_notebook_file,
     upload_notebook_file,
-    create_job_from_notebook,
 )
 
-__all__ = [
-    "create_notebook_file",
-    "upload_notebook_file",
-    "create_job_from_notebook",
-]
+__all__ = ["create_notebook_file", "upload_notebook_file"]
