+++ src/api/tools/notebooks/notebooks.py
@@ -370,86 +370,3 @@ async def upload_notebook_file(
             "fileSize": os.path.getsize(local_path),
         },
     }
-
-
-async def create_job_from_notebook(
-    ctx: Context,
-    name: str,
-    notebook_path: str,
-    mode: str = "Once",
-    execution_interval_in_minutes: Optional[int] = None,
-) -> dict:
-    """
-    Create a scheduled job to run a notebook (uploaded to shared space).
-
-    Args:
-        ctx: Context object
-        name: Name of the job
-        notebook_path: Remote path to the shared notebook file
-        mode: Job mode (options: "Once", "Recurring")
-        execution_interval_in_minutes: Optional interval in minutes for recurring jobs
-
-    Returns:
-        Dict with job creation result
-    """
-
-    logger.info("Creating job from notebook")
-
-    settings = config.get_settings()
-    start_time = time.time()
-    user_id = config.get_user_id()
-    try:
-        workspace_manager = s2.manage_workspaces(
-            access_token=get_access_token(),
-            base_url=settings.s2_api_base_url,
-            organization_id=get_org_id(),
-        )
-        org = workspace_manager.organization
-        jobs_manager = org.jobs
-
-        job_obj = jobs_manager.schedule(
-            notebook_path=notebook_path,
-            name=name,
-            mode=s2.management.job.Mode(mode),
-            create_snapshot=True,
-            execution_interval_in_minutes=execution_interval_in_minutes,
-        )
-
-        settings.analytics_manager.track_event(
-            user_id,
-            "tool_calling",
-            {
-                "name": "create_job_from_notebook",
-                "job_name": name,
-                "notebook_path": notebook_path,
-            },
-        )
-        execution_time = (time.time() - start_time) * 1000
-        return {
-            "status": "success",
-            "message": f"Job '{name}' created successfully.",
-            "data": {
-                "jobID": job_obj.job_id,
-                "name": job_obj.name,
-                "description": job_obj.description,
-                "completedExecutionsCount": job_obj.completed_executions_count,
-                "schedule": {
-                    "mode": job_obj.schedule.mode.value,
-                    "executionIntervalInMinutes": job_obj.schedule.execution_interval_in_minutes,
-                },
-                "createdAt": job_obj.created_at,
-                "terminatedAt": job_obj.terminated_at,
-            },
-            "metadata": {
-                "executionTimeMs": round(execution_time, 2),
-                "timestamp": datetime.now(timezone.utc).isoformat(),
-            },
-        }
-    except Exception as e:
-        logger.error(f"Error creating job: {str(e)}")
-        return {
-            "status": "error",
-            "message": f"Failed to create job: {str(e)}",
-            "errorCode": "JOB_CREATION_FAILED",
-            "errorDetails": {"exception_type": type(e).__name__},
-        }
