+++ src/api/tools/notebooks/utils.py
@@ -8,9 +8,11 @@ from typing import Any, Dict, Optional
 
 import nbformat as nbf
 import nbformat.v4 as nbfv4
+
 from src.config import config
 from src.api.common import get_access_token, get_org_id
 from src.logger import get_logger
+
 
 # Set up logger for this module
 logger = get_logger()
