+++ src/api/tools/tools.py
@@ -16,8 +16,8 @@ from src.api.tools.user import get_user_info
 from src.api.tools.notebooks import (
     create_notebook_file,
     upload_notebook_file,
-    create_job_from_notebook,
 )
+from src.api.tools.jobs import create_job_from_notebook, delete_job
 from src.api.tools.organization import (
     organization_info,
     choose_organization,
@@ -41,6 +41,7 @@ tools_definition = [
     {"func": create_notebook_file},
     {"func": upload_notebook_file},
     {"func": create_job_from_notebook},
+    {"func": delete_job},
 ]
 
 # Export the tools
